[FIX] Change property to empty value
This fixes the case when you clear the value of a property which
resulted into setting it to "0".
Now "null" will be used which set the property to the default value.

However this might not be the expected result (e.g. when the default
value is a string, which makes it impossible to clear it).

A way to explicitly reset a property value might be a better solution.