[FEATURE] Bindings tab: more detailed information of underlying data
A property can have multiple bindings that are combined via a formatter.
With this change all relevant bindings are listed and linked to the model.
This is enabled for aggregations as well.
Furthermore the naming of the model entry was unified, the binding
mode is not part of the model description anymore but part of the
binding information.