fix: fixed UI5 version detection
Problem: When the inspected app is built the UI5 Tooling, the version
which we take from the "sap.ui.getVersionInfo" API is incorrect.
This results in some unwanted behavior in the "Event Registry" tab of
the inspector.

Solution: We take the framework version from the framework's core
configuration which always returns the correct data.