fix: replaced root font size metrics
Problem: In Safari some styles were not applied correctly because of
the root font size metrics. This was due to the fact that the root font
size was set to 62.5% of the browser's default font size.

Solution: The root font size is set to 16px, which is the default font
size in most browsers. This allows the styles to be applied correctly.