feat: copy control to console (#215)
A click on the id of a control in the properties
tab or on the button 'Copy to console' in the
Action tab will create a temp global variable
starting with 'ui5$' and a number e.g. 'ui5$0'
that refers to the control. All controls
will be collected in the variable 'ui5$temp' for
convenience.